Add struct to be returned back with an enclave Report (#3509)
When an enclave generates a `Report` it has a need to provide a nonce a
public key, and custom identity. The fog enclave would use the custom
identity to store the ingress key.
Previously the nonce was provided to the quoting enclave during quote
genration. This would then be returned in the quoting enclave's
`Report`. The public key and custom identity was stored in the app
enclave's `ReportBody.report_data`.

For DCAP, the quoting enclave's no longer have a means to store a nonce.
The app enclave's `ReportBody.report_data` is only 64 bytes and is thus
at capacity with no room for the nonce. To work with the nonce, the DCAP
app enclave's will store a hash over the nonce, public key, and custom
identity. When a DCAP app enclave is queried for a `Report` it will
return the `Report` as well as the structure holding the nonce, public
key and custom identifier. During quote verification this structure will
be hashed and compared to the nonce stored in the quote's app enclave.

This change brings in the containing struct `EncalveReportDataContents`
and the quote verification logic needed for it.