Use a MRSIGNER identity for watcher.
Previously watcher didn't verify VerificationReports against a known
identity. This was because watcher isn't trying to securely communicate
with an enclave, it's leveraging the common attestation API to get
access to the VerificationReports. The VerificationReports are stored
for historical proof of the enclaves. The new DCAP verifier requires an
identity in order to succeed. Now the watcher will provide a MRSIGNER
identity with a 0 SVN so that any signed consensus node will succeed in
verifying, providing watcher with the verification report.