Use `EnclaveReportDataContents` for quoting
Previously a `QuoteNonce` was passed back from generating an enclave
report and that nonce was used in generating a quote as well as in
verifying the quote. Now an `EnclaveReportDataContents` is passed back
from an enclave report and used to quote and verify the quote. The
reason for this is to support storing more data than the 64 byte
`ReportData` member of the Quote.