Add `Digestible` impl to `EnclaveReportDataContents`
The `EnclaveReportDataContents` is a member of `DcapEvidence` which will
need to be part of `BlockMetadataContents` which derives `Digestible`.
In order to support this future need a `Digestible` implementation has
been added to the `EnclaveReportDataContents`.

Also add `EnclaveReportDataContents` to the external protobufs. To
accomodate the need of it being part of the `BlockMetadataContents` in
the protobufs.