Add UnverifiedReport state for attestation (#3516)
* Add UnverifiedReport state for attestation

A new state `UnverifiedReport` as been added to the attest key exchange
state machine. The `UnverifiedReport` allows clients like watcher to
get the attestation report from the key exchange handshake without
requiring the report to pass verification. This is useful when the
client doesn't know or need to know the identity of the enclave.

* Remove ready state for the `UnverifiedReport` state

Previously the `UnverifiedReport` returned a `Ready` state which would
allow subsequent communication. Now the `UnverifiedReport` returns a
`Terminated` state. This is meant to prevent a client from accidentally
talking to an unknown enclave, thinking it's trusted, by using an
`UnverifiedReport` handshake.