Fix mobilecoind tests
Previously some mobilecoind tests were trying to add blocks with
`BlockVersion::THREE` after adding blocks with `BLOCK_VERSION`.
`BLOCK_VERSION` is `BlockVersion::MAX` which has been updated to `FOUR`.

Ledger asserts that blocks are the same or increasing in version. With a
new max of FOUR, this resulted in the tests trying to add a block
version THREE after blocks of version FOUR.