Add prost implementations for `Quote3` and `Collateral`
Add prost implementations for `Quote3` and `Collateral`. The
implementations are done in this repository instead of
`mc-sgx-dcap-types` to prevent having to publish proto files from the
`mc-stx-***` crates.

The `prost::Message` implementations use `prost-build` and a `*.proto`
file, which deviates from other `prost::Message` implementations in the
repository. The reasons for this deviation are:
- Prevent changing `Quote3` and `Collateral` to use `Option<T>`
  fields. `Option<T>` or defaults need to be implemented to work
  directly with `::prost::Message`.
- Ensure that the `::prost::Message` implementations use the same tags
  as the gRPCio protobuf implementations.