Moving the `DcapEvidence` protobufs to `mc-api` (#3579)
Previously the protobufs for `DcapEvidence` and its sub-structs were in
`mc-attest-api`. Due to dependencies in `mc-attest-api` having the
`DcapEvidence` type defined there would result in the wasm builds failing
once types in `mc-api` need to depend on `mc-attest-api`. To prevent
this future failure the `DcapEvidence` is brought into the `mc-api`
crate. This is the crate that houses the `VerificationReport` which the
`DcapEvidence` is meant to replace.