Change type of `mc-attest-verifier::Error::Verification` (#3583)
Previously the `mc-attest-verifier::Error::Verification` type wrapped a
`VerificationReportData`. Now it wraps a `String` type. This allows for
a more generic interface to provide back the error from attempting to
verify the legacy `VerificationReportData` or the newer `DcapEvidence`.

Verification of the newer `DcapEvidence` provides a string output that
more clearly states where verification failed and as such passing back
the evidence itself isn't very beneficial.