Fix an error code returned by mobilecoind
When checking that a b58 address is valid, if the code cannot be
parsed, that should be considered an invalid argument error rather
than an internal error.

Improving this to return the right error code helps users of this API
handle errors correctly themselves, since internal errors are often
considered retriable but an invalid argument is not.