Remove timestamp age out in validity check.
Previously a timestamp in SCP would be marked invalid if it exceeded 30
seconds. This posed a problem if multiple nodes were flooded with
transactions that exceeded their capacity, and the nodes were stalled
enough such that it took more than 30 seconds to get the value +
timestamp to a quorum. Now timestamps are only ensured to not be in the
future and to be increasing from the last block's timestamp.