Fix ledger stores not updating latest block info
In a a7b70f620bc4c8959629978247002d100af85b45 the ledger stores were
optimized to fetch only the blocks they needed. This caused an error
during collation where the minimum of the known blocks are past back to
clients. This resulted in only being able to provide the blocks from the
lowest shard to the clients. Now shards will continue to poll for the
latest block allowing for all shards to be accounted for when collating
the response.