Move sim certificates to build directory
Previously the sim enclave signing certificates were in the source
directory and used a random key. This caused issues with incremental
builds and at timing issues where the enclaves were in the process of
creating a certificate while the untrusted code was too. This resulted
in attestation errors due to signature mismatches, which would force
someone to rebuild again. Now the sim signing certificates are in the
build directory and use a common default seed.