Add support for setting tx private key in tx builder, expose in mobilecoind api (#3879)
* Add support for setting tx private key in tx builder, expose in mobilecoind api

This can be used to create a relatively easy to use idempotence strategy
for mobilecoind payments.

From a vendor or an exchange's point of view, it's really nice if
you can have an "idempotence key" in an API that sends money somewhere.
Here idempotence means, if I make this request, and retry in case of connection
failures, there's no way that it can lead to a payment being sent twice.

For example, in mobilecoind, there is a `SendPayment` API which is meant to
be the easiest way to send money to a public address. However, there is no
idempotence key on this API, and so if I make the request, then due to
connectivity issues, the response is not delivered, but the request was
handled successfully, a retry will cause a new payment to be sent, with new
UTXOs selected. So in realistic deployment environments, this API cannot
actually be used safely by an exchange or a vendor. Some fraction of the
time, random networking issues will cause payments to be delivered twice,
and loss of funds.

It's best for the idempotence mechanism to be grounded in blockchain rules,
because local wallet state is comparatively ephemeral.

The mobilecoin blockchain does not have a built-in notion of idempotence keys,
but it does have two points of exclusion for transactions -- a UTXO cannot be
spent twice, due to the key image mechanism, and a TxOut public key is unique
across the whole chain.

One way that an idempotence mechanism can be created is by fixing the
tx private key of (one or more) TxOut's in a transaction that you create.
The tx private key is the only random element used when the TxOut public key,
target key, and amount commitment are generated for a given recipient, token id,
and value, so fixing the recipient, amount, and tx private key, fixes these
fields of the TxOut, in particular the public key.

From an exchange or vendor's point of view, fixing the tx private key has
additional benefits -- if you know the tx private key, amount, and recipient,
you can prove the amount and recipient of a TxOut that you sent, even if the
recipient themselves does not cooperate. This can be helpful to resolve disputes.
Alternative existing mechanisms like shared secret and confirmation number
don't accomplish this, because they can only be checked by the recipient,
whereas this tx private key mechanism can be checked by anyone.

Existing wallets do not expose the tx private key, it is generally discarded instead.
This is perhaps very good from a privacy point of view, but it basically gives up
on robust dispute resolution.

This PR gives an exchange or vendor a relatively easy way to get both things.
Such a user could decide to derive the tx_private_key of a TxOut, for instance,
by a hash of the payment id or withdrawal id, together with a 32 byte secret.

As an example, the private spend key could be the secret used in this hash, however,
that has some unexpected consequences like, rotating your private key could
cause a loss of idempotence. As an alternative, a user could even choose to use
an entirely new secret here which is not derived from the mnemonic, so
that they can still change their mnemonic without risking loss of idempotence
(and loss of funds).

Instead of mandating any such mechanism, this PR allows the application developer to
decide exactly how this should be done. The new `tx_private_key` field of an `Outlay`
is optional, but when specified, is expected to be a 32-byte hash.
When used this way, a call to the `SendPayment` API can become idempotent.

---

Note that, this is an alternative to an approach developed some time ago in the
mobile SDKs, when mobile wallets needed to establish idempotence for payments.
The alternative approach is to seed the RNG used with the transaction builder,
in order to control the tx private keys. This seed was often a hash using the
payment id.

The main drawback of this is that it is fairly brittle.

* Invoking transaction builder functions in a different order can break idempotence,
  so changing code that uses it may become hazardous.
* Upgrading the seeded CSPRNG algortihm can break idempotence, and so becomes
  hazardous.

Additionally, from the user point of view, it only accomplishes one of my goals.
It is a strategy for controlling the tx private key, but it doesn't help me
learn the tx private key, and none of the existing APIs return that value.

I believe that this approach is the simplest change that accomplishes both goals,
and is backwards compatible from an API perspective.
Application developers are free to use either approach to establish and maintain
idempotence.

* Update mobilecoind/src/conversions.rs

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* Update mobilecoind/src/conversions.rs

Co-authored-by: Nick Santana <nick@mobilecoin.com>

* fixup previous

* fixup previous after rebase

* fix more clippies

---------

Co-authored-by: Nick Santana <nick@mobilecoin.com>