Fix fog ledger router returning incorrect block number
Previously when a fog ledger store was spun up prior to the block chain
having blocks in its block range it would return back 0 for the block
number. This resulted in the collated response having 0 for the total
number of blocks. Now stores will always respond back with the number of
blocks their block provider has.