Fix fog ledger router returning incorrect block number
Previously when a fog ledger store was spun up prior to the block chain
having blocks in its block range it would return back 0 for the block
number. This resulted in the collated response having 0 for the total
number of blocks. Now stores will respond back with the range of
blocks that they have processed and the collation logic will combine the
overlapping ranges to come up with the highest available block count.